Factor out typechecking logic and add a prompt about Typechecking
